virt_mshv: organize SNP support under x86_64::snp
Move SNP protocol helpers, launch and GHCB handlers, state ownership, and focused tests out of the shared x86_64 backend module. Group partition and VP SNP state behind backend-local optional state, use that state as the runtime isolation gate, and remove the mid-file test-module lint suppression.

Review feedback: chris-oo/openvmm#9 comments 3768723726, 3768725181, 3768734416, 3768765849, 3768794715
Final stack placement: create SnpPartitionState with launch_state in zknyppqnzvrq, add sev_features in rrrqxyrtmvus, and fold SnpVpState/GHCB ownership into xllzxnulpxrs. Keep the remaining file-only organization as the final commit after pprqxmxqtskz and wqtxvxrmmmly, with the state types in x86_64/mod.rs until that move.

Copilot-Session: eb0c5e4d-729c-4d16-b625-b44e2a97a2fb
